+++ payload/apps/docs/next.config.ts
@@ -1,9 +1,0 @@
-import type { NextConfig } from "next";
-
-const nextConfig: NextConfig = {
-  poweredByHeader: false,
-  output: "standalone",
-  reactStrictMode: true,
-};
-
-export default nextConfig;